chore: split wallet stuff into its' own classes
