+++ libs/wallet/src/connectors/embedded-connector/index.ts
@@ -1,0 +1,2 @@
+export * from './connector';
+export * from './quick-start-connector';
