+++ libs/wallet/src/connectors/index.ts
@@ -1,9 +1,6 @@
 export { InjectedConnector } from './injected-connector';
 export { SnapConnector } from './snap-connector';
-export {
-  InBrowserConnector,
-  QuickStartConnector,
-} from './in-browser-connector';
+export { InBrowserConnector, QuickStartConnector } from './embedded-connector';
 export { JsonRpcConnector } from './json-rpc-connector';
 export { ViewPartyConnector } from './view-party-connector';
 export { MockConnector, mockKeys } from './mock-connector';
